Revert "migrate near/far point data config from TrackGui to TrackEndPointGui"
This reverts commit 3e28247220c3c846b714a6d123e36ad635309863.
