+++ src/main/java/cam72cam/immersiverailroading/library/GuiText.java
@@ -19,10 +19,8 @@ public enum GuiText {
 	LABEL_PITCH_SETTING_COMMON("label.pitch_setting_common"),
 	LABEL_PITCH_SETTING_MIX("label.pitch_setting_mix"),
 	LABEL_PITCH_SETTING_DEGREE("label.pitch_setting_degree"),
-	LABEL_PITCH_SETTING("label.pitch_setting"),
-	LABEL_POS_YAW_TYPE("label.pos_yaw_type"),
-	LABEL_NEAR("label.near"),
-	LABEL_FAR("label.far"),
+	LABEL_NEAR_PITCH_SETTING("label.near_pitch_setting"),
+	LABEL_FAR_PITCH_SETTING("label.far_pitch_setting"),
 	SLIDER_ZOOM("slider.zoom"), //Need colon
 	SELECTOR_PAGE("selector.page"),
 	SELECTOR_TYPE("selector.type"),
@@ -83,7 +81,6 @@ public enum GuiText {
 
 	TRACK_EXTRA_TO_MAIN("track.extra_trackgui"),
 	TRACK_MAIN_TO_EXTRA("track.extragui"),
-	TRACK_MAIN_TO_END_POINT("track.endpointgui"),
 	
 	LOCO_WORKS("loco.works"),
 //	LOCO_HORSE_POWER("loco.horse_power"),
