+++ src/main/resources/assets/immersiverailroading/lang/en_us.lang
@@ -72,19 +72,17 @@ gui.immersiverailroading:label.length=Length:
 gui.immersiverailroading:label.no_length=Length Not Used
 gui.immersiverailroading:label.radius=Radius:
 gui.immersiverailroading:label.length_radius=Length/Radius:
-gui.immersiverailroading:label.transition_radius=Transition Radius:
+gui.immersiverailroading:label.transition_radius=Near/Far Radius:
 gui.immersiverailroading:label.swap_radius=Swap Radius
-gui.immersiverailroading:label.height_offset=Height Offset
-gui.immersiverailroading:label.reset_height_offset=Reset Height Offset
-gui.immersiverailroading:label.pitch=Pitch
+gui.immersiverailroading:label.height_offset=Near/Far Y Offset
+gui.immersiverailroading:label.reset_height_offset=Reset Y Offset
+gui.immersiverailroading:label.pitch=Near/Far Pitch
 gui.immersiverailroading:label.reset_pitch=Reset Pitch
 gui.immersiverailroading:label.pitch_setting_common=Common Mode
 gui.immersiverailroading:label.pitch_setting_mix=Mix Mode
 gui.immersiverailroading:label.pitch_setting_degree=Degree Mode
-gui.immersiverailroading:label.pitch_setting=Pitch Calculating Method
-gui.immersiverailroading:label.pos_yaw_type=Direction Type
-gui.immersiverailroading:label.near=Near
-gui.immersiverailroading:label.far=Far
+gui.immersiverailroading:label.near_pitch_setting=Near Pitch Calculating Method
+gui.immersiverailroading:label.far_pitch_setting=Far Pitch Calculating Method
 gui.immersiverailroading:slider.zoom=Zoom:
 gui.immersiverailroading:selector.page=Page %s of %s
 gui.immersiverailroading:selector.type=Type: %s
@@ -156,7 +154,6 @@ gui.immersiverailroading:track.extra_edit_left=Edit Left
 gui.immersiverailroading:track.extra_edit_right=Edit Right
 gui.immersiverailroading:track.extra_trackgui=Basic
 gui.immersiverailroading:track.extragui=Advanced
-gui.immersiverailroading:track.endpointgui=End Point
 
 gui.immersiverailroading:loco.works=Works:          %s
 gui.immersiverailroading:loco.power=Power:          %s
