+++ src/main/resources/assets/immersiverailroading/lang/zh_cn.lang
@@ -72,19 +72,17 @@ gui.immersiverailroading:label.length=长度：
 gui.immersiverailroading:label.no_length=长度未使用
 gui.immersiverailroading:label.radius=半径：
 gui.immersiverailroading:label.length_radius=长度/半径：
-gui.immersiverailroading:label.transition_radius=缓和曲线半径：
+gui.immersiverailroading:label.transition_radius=近/远端半径：
 gui.immersiverailroading:label.swap_radius=交换半径
-gui.immersiverailroading:label.height_offset=高度偏移
+gui.immersiverailroading:label.height_offset=近/远端高度偏移
 gui.immersiverailroading:label.reset_height_offset=重置高度偏移
-gui.immersiverailroading:label.pitch=坡度
+gui.immersiverailroading:label.pitch=近/远端坡度
 gui.immersiverailroading:label.reset_pitch=重置坡度
 gui.immersiverailroading:label.pitch_setting_common=普通模式
 gui.immersiverailroading:label.pitch_setting_mix=混合模式
 gui.immersiverailroading:label.pitch_setting_degree=角度模式
-gui.immersiverailroading:label.pitch_setting=坡度计算方式
-gui.immersiverailroading:label.pos_yaw_type=朝向类型
-gui.immersiverailroading:label.near=近端
-gui.immersiverailroading:label.far=远端
+gui.immersiverailroading:label.near_pitch_setting=近端坡度计算方式
+gui.immersiverailroading:label.far_pitch_setting=远端坡度计算方式
 gui.immersiverailroading:slider.zoom=缩放比例：
 gui.immersiverailroading:selector.page=第 %s 页，共 %s 页
 gui.immersiverailroading:selector.type=类型：%s
@@ -156,7 +154,6 @@ gui.immersiverailroading:track.extra_edit_left=修改左侧
 gui.immersiverailroading:track.extra_edit_right=修改右侧
 gui.immersiverailroading:track.extra_trackgui=主界面
 gui.immersiverailroading:track.extragui=附加设置界面
-gui.immersiverailroading:track.endpointgui=端点设置界面
 
 gui.immersiverailroading:loco.works=服务公司：%s
 gui.immersiverailroading:loco.power=功率：%s
